services/automation-agent/requirements.txt
# Core automation libraries
agentql>=1.0.0
playwright>=1.40.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0

# Anti-detection and stealth
fake-useragent>=1.4.0
requests-html>=0.10.0
httpx[http2]>=0.25.0
curl-cffi>=0.5.0

# Proxy management
aiohttp-socks>=0.8.0
python-socks>=2.4.0

# Data handling and configuration
pydantic>=2.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
loguru>=0.7.0

# Pokemon/retail specific
beautifulsoup4>=4.12.0
lxml>=4.9.0
selectolax>=0.3.0

# Async and concurrency
asyncio-throttle>=1.0.2
aiofiles>=23.2.0

# Monitoring and alerts
discord.py>=2.3.0
twilio>=8.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-playwright>=0.4.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0